feat: make create question a route
