+++ src/app/pages/teacher/components/home/home.component.html
@@ -1,38 +1,7 @@
-<main class="">
-  <div class="flex justify-between items-center mb-8">
-    <div></div>
-    <div class="flex gap-2">
-      <button
-        class="text-[#175CD3] font-WorkSans font-semibold text-sm mt-4 bg-[#D1E9FF] px-2 py-1 rounded-sm transition-all ease-in 1s"
-      >
-        Question
-      </button>
-      <button
-        class="text-[#404653] font-WorkSans font-semibold text-sm mt-4 px-2 py-1 rounded-sm transition-all ease-in 1s"
-      >
-        Settting
-      </button>
-    </div>
+<section class="">
+  @if(isCreateQuestionsPage()) {
 
-    <div class="flex gap-2">
-      <button
-        class="text-black border-gray-400 text-xs mt-4 hover:bg-gray-100 bg-white border px-3 py-1 rounded-md transition-all ease-in 1s"
-      >
-        Preview
-      </button>
-      <button
-        type="button"
-        (click)="sendQuestions()"
-        class="text-white text-sm mt-4 hover:bg-blue-500 bg-[#175CD3] px-3 py-1 rounded-md transition-all ease-in 1s"
-      >
-        Send
-      </button>
-    </div>
-  </div>
-  @if(toggledCreateQuestion()){
-  <div class="md:w-[60%] md:mx-auto">
-    <app-form (outputQuestions)="handleQuestions($event)"/>
-  </div>
+  <router-outlet></router-outlet>
 
   } @else {
 
@@ -47,7 +16,7 @@
 
       <button
         type="button"
-        (click)="toggleCreateQuestion()"
+        (click)="createQuestions()"
         class="bg-[#1849A9] text-white py-4 px-6 flex items-center justify-center gap-3 rounded-md text-sm"
       >
         <svg
@@ -69,4 +38,4 @@
     </div>
   </div>
   }
-</main>
+</section>
