+++ src/app/pages/teacher/components/home/home.component.ts
@@ -1,77 +1,21 @@
-import { Component, inject, signal } from '@angular/core';
-import { FormComponent } from '../form/form.component';
-import { ExamQuestion, Question } from '../../interfaces/question.interface';
-import { QuestionsService } from '../../services/questions/questions.service';
-import { ToasterService } from '../../../../shared/components/toaster/services/toaster.service';
-import { IResponse } from '../../../../interfaces/response.interface';
+import { Component, inject } from '@angular/core';
+import { Router, RouterLink, RouterOutlet } from '@angular/router';
 
 @Component({
   selector: 'app-home',
   standalone: true,
-  imports: [FormComponent],
+  imports: [RouterOutlet, RouterLink],
   templateUrl: './home.component.html',
   styleUrl: './home.component.scss',
 })
 export class HomeComponent {
-  isLoading: boolean = false;
-  toggledCreateQuestion = signal<boolean>(false);
-  currentQuestions = signal<Question[]>([]);
+  private _router = inject(Router);
 
-  private _questionsService = inject(QuestionsService);
-  private _toaster = inject(ToasterService);
-
-  toggleCreateQuestion() {
-    this.toggledCreateQuestion.update((prev) => !prev);
+  isCreateQuestionsPage() {
+    return this._router.url === '/teacher/home/create-questions';
   }
 
-  handleQuestions(questions: Question[]) {
-    this.currentQuestions.set(questions);
-  }
-
-  sendQuestions() {
-    const data: ExamQuestion = {
-      questionTitle: 'Software Testing',
-      questionInstruction: 'Mid-semester question',
-      questionStartTime: '2024-06-25T10:56:54.730Z',
-      questionEndTime: '2024-06-26T10:56:54.730Z',
-      question: this.currentQuestions(),
-      questionReceivers: [
-        'nusetorsetsofia102@gmail.com',
-        'sandro@gmail.com',
-        'kwame@gmail.com',
-      ],
-    };
-
-    this._questionsService.createQuestion(data).subscribe({
-      next: (response) => {
-        this.responseHandler(response);
-      },
-      error: (error) => {
-        this.errorHandler(error);
-      },
-    });
-  }
-
-  responseHandler(response: IResponse) {
-    this.isLoading = false;
-
-    if (response.status === 200) {
-      this.showToast(response.message, 'success');
-    } else {
-      this.showToast(response.message, 'error');
-    }
-  }
-
-  errorHandler(error: Error) {
-    this.isLoading = false;
-    this.showToast(error.message, 'error');
-  }
-
-  private showToast(message: string, type: 'error' | 'success') {
-    this._toaster.showToast({
-      message,
-      type,
-      duration: 3000,
-    });
+  createQuestions() {
+    this._router.navigate(['teacher', 'home', 'create-questions']);
   }
 }
