+++ src/app/pages/teacher/teacher.routes.ts
@@ -1,6 +1,8 @@
 import { Routes } from '@angular/router';
 import { HomeComponent } from './components/home/home.component';
 import { AnalyticsComponent } from './components/analytics/analytics.component';
+import { FormComponent } from './components/form/form.component';
+import { SetQuestionsComponent } from './components/set-questions/set-questions.component';
 
 export const routes: Routes = [
   {
@@ -11,9 +13,15 @@ export const routes: Routes = [
   {
     path: 'home',
     component: HomeComponent,
+    children: [
+      {
+        path: 'create-questions',
+        component: SetQuestionsComponent,
+      },
+    ]
   },
   {
     path: 'analytics',
     component: AnalyticsComponent,
-  },
+  }
 ];
